fixes #67 regenerate petstore project in both openapi and swagger for 1.5.14

+++ rest/openapi/petstore/src/main/java/com/networknt/petstore/model/Pet.java
@@ -8,7 +8,7 @@ public class Pet {
     
     private String name;
     
-    private Integer id;
+    private java.lang.Long id;
     
     private String tag;
     
@@ -30,11 +30,11 @@ public class Pet {
     
     
     @JsonProperty("id")
-    public Integer getId() {
+    public java.lang.Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(java.lang.Long id) {
         this.id = id;
     }
